moves test file to mpl_toolkits/tests/
