+++ lib/matplotlib/tests/test_axes.py
@@ -34,7 +34,6 @@ from numpy.testing import (
     assert_allclose, assert_array_equal, assert_array_almost_equal)
 from matplotlib import rc_context
 from matplotlib.cbook import MatplotlibDeprecationWarning
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 # Note: Some test cases are run twice: once normally and once with labeled data
 #       These two must be defined in the same test function or need to have
@@ -6825,31 +6824,4 @@ def test_ylabel_ha_with_position(ha):
     ax.set_ylabel("test", y=1, ha=ha)
     ax.yaxis.set_label_position("right")
     assert ax.yaxis.get_label().get_ha() == ha
-    
-@image_comparison(['second_axis_with_colorbar.png'])
-def test_second_axis_with_colorbar():
-	X = np.array([1,2,3,4])
-	X2 = np.array([1,2,3,4])
-	Y = np.cos(X*20)
-	Z = np.sin(X*20)
 
-	fig = plt.figure()
-	ax1 = plt.subplot(111)
-	ax2 = ax1.twiny()
-
-	ax1.minorticks_off()
-
-	SC = ax1.scatter(X, Y, c=Z)
-	ax1.set_xlabel("Original x-axis")
-
-	ax2.set_xlim(ax1.get_xlim())
-	ax2.set_xticks(X)
-	ax2.set_xticklabels(X2)
-	ax2.set_xlabel("Second x-axis")
-
-	# Colorbar.
-	the_divider = make_axes_locatable(ax1)
-	color_axis = the_divider.append_axes("right", size="2%", pad=0.1)
-	cbar = plt.colorbar(SC, cax=color_axis)
-	cbar.set_label('B', fontsize=10, labelpad=4, y=0.5)
-	cbar.ax.tick_params(labelsize=10)
